Feature/models

+++ backend/Backend.Domain/Entities/MovieActor.cs
@@ -4,9 +4,11 @@ namespace Backend.Domain.Entities;
 
 public class MovieActor : IEntity
 {
-    public long Id { get; set; }
-    public long MovieId { get; set; }
-    public long ActorId { get; set; }
+    public int Id { get; set; }
+
+    public int MovieId { get; set; }
+
+    public int ActorId { get; set; }
 
     public Movie Movie { get; set; } = null!;
     public Actor Actor { get; set; } = null!;
