+++ backend/Backend.Domain/Entities/MovieGenre.cs
@@ -4,9 +4,11 @@ namespace Backend.Domain.Entities;
 
 public class MovieGenre : IEntity
 {
-    public long Id { get; set; }
-    public long MovieId { get; set; }
-    public long GenreId { get; set; }
+    public int Id { get; set; }
+
+    public int MovieId { get; set; }
+
+    public int GenreId { get; set; }
 
     public Movie Movie { get; set; } = null!;
     public Genre Genre { get; set; } = null!;
